+++ src/components/IdeaForm.vue
@@ -14,7 +14,7 @@
         <q-chips-input v-model="tags" float-label="Tags" dark />
       </div>
       <div class="col-12">
-        <q-chips-input v-model="bibleRefs" float-label="Bible Refs" dark />
+        <q-chips-input v-model="bibleRefs" float-label="Bible Refs" color="secondary" dark />
       </div>
       <div class="col-12">
         <q-input v-model="notes" type="textarea" :max-height="100" :min-rows="2" float-label="Notes" dark />
@@ -34,8 +34,8 @@
         <q-input type="number" v-model="location.end" v-if="locationType !== 'None'" :float-label="locationType + ' End'" dark frame-color="secondary" />
       </div>
       <div class="col-12">
-        <q-btn v-if="formType === 'Add'" color="primary" @click="addIdea">Add</q-btn>
-        <q-btn v-if="formType === 'Edit'" color="primary" @click="updateIdea">Update</q-btn>
+        <q-btn v-if="formType === 'Add'" color="primary" @click="addIdea" class="on-left">Add</q-btn>
+        <q-btn v-if="formType === 'Edit'" color="primary" @click="updateIdea" class="on-left">Update</q-btn>
         <q-btn v-if="formType === 'Edit'" color="negative" @click="removeIdea">Remove</q-btn>
       </div>
     </div>