+++ src/pages/Document.vue
@@ -2,7 +2,11 @@
   <q-page padding>
     <div class="row gutter-md items-center">
       <div class="col-xs-12 justify-center">
-        <pdf v-if="document.fileType === 'application/pdf' && fileURL !== ''" :src="fileURL" />
+        <q-btn v-if="fileTypes.includes(document.fileType) && fileURL !== ''" @click.native="wordRead = true">read</q-btn>
+        <q-btn v-if="document.fileType === 'application/pdf' && fileURL !== ''" @click.native="pdfRead = true">read</q-btn>
+        <q-btn v-if="document.fileType === 'application/pdf' && fileURL !== ''" @click.native="$refs.pdf.print()">print</q-btn>
+        <!-- <q-btn v-if="document.fileType === 'text/html' && fileURL !== ''" @click.native="htmlRead = true">read</q-btn>
+        <q-btn v-if="document.fileType === 'text/csv' && fileURL !== ''" @click.native="textRead = true">read</q-btn> -->
       </div>
       <div class="col-xs-12">
         <span class="float-right" v-if="this.$selectedTopic.get()">
@@ -24,6 +28,9 @@
             <q-chips-input v-model="userData.tags" float-label="Tags" dark />
           </div>
           <div class="col-12">
+            <q-input v-model="document.citation" type="textarea" :max-height="100" :min-rows="2" float-label="Citation" dark />
+          </div>
+          <div class="col-12">
             <q-btn color="primary" @click="update">Update</q-btn>
             <q-btn color="negative" class="float-right" @click="remove">Remove</q-btn>
           </div>
@@ -33,6 +40,36 @@
         <media-notes :user-notes="userData.notes" :update="updateNotes" :mediaid="id" media-type="document"></media-notes>
       </div>
     </div>
+    <q-modal v-if="document.fileType === 'application/pdf' && fileURL !== ''" v-model="pdfRead" maximized>
+      <q-btn class="overlay-button fixed z-max" style="top: 20px; right: 20px;" color="primary" icon="fas fa-times" @click.native="pdfRead = false" />
+      <q-btn color="primary" size="xl" class="overlay-button fixed z-max" style="left: 0px; top: 50%; transform: translate(50%, -50%); -webkit-transform: translate(50%, -50%);" icon="fas fa-chevron-left" @click="prevPage" />
+      <q-btn color="primary" size="xl" class="overlay-button fixed z-max" style="right: 0px; top: 50%; transform: translate(-50%, -50%); -webkit-transform: translate(-50%, -50%);" icon="fas fa-chevron-right" @click="nextPage" />
+      <!-- <q-btn-group class="fixed z-max" style="bottom: 0px; left: 50%; transform: translate(-50%, -50%); -webkit-transform: translate(-50%, -50%);">
+        <q-btn color="primary" size="xl" class="overlay-button" icon="fas fa-chevron-left" @click="prevPage" />
+        <q-btn color="primary" size="xl" class="overlay-button" icon="fas fa-chevron-right" @click="nextPage" />
+      </q-btn-group> -->
+      <pdf
+        :src="fileURL"
+        :page="currentPage"
+        ref="pdf"
+        @num-pages="pageCount = $event"
+        @page-loaded="currentPage = $event"
+        @link-clicked="currentPage = $event"
+        v-touch-swipe.horizontal="handleSwipe"
+      />
+    </q-modal>
+    <q-modal v-if="fileTypes.includes(document.fileType) && fileURL !== ''" v-model="wordRead" maximized>
+      <q-btn class="overlay-button fixed z-max" style="top: 20px; right: 20px;" color="primary" icon="fas fa-times" @click.native="wordRead = false" />
+      <iframe :src="'https://view.officeapps.live.com/op/view.aspx?src=https://storage.googleapis.com/notes-and-quotes-977a3.appspot.com/documents/' + id" width='100%' height='100%' frameborder='0'>This is an embedded <a target='_blank' href='http://office.com'>Microsoft Office</a> document, powered by <a target='_blank' href='http://office.com/webapps'>Office Online</a>.</iframe>
+      <!-- <iframe :src="'http://docs.google.com/gview?url=' + fileURL + '&embedded=true'" width='100%' height='100%' frameborder='0'>This is an embedded <a target='_blank' href='http://office.com'>Microsoft Office</a> document, powered by <a target='_blank' href='http://office.com/webapps'>Office Online</a>.</iframe> -->
+    </q-modal>
+    <!-- <q-modal v-if="document.fileType === 'text/html' && fileURL !== ''" v-model="htmlRead" maximized>
+      <q-btn class="overlay-button fixed z-max" style="top: 20px; right: 20px;" color="primary" icon="fas fa-times" @click.native="htmlRead = false" />
+    </q-modal>
+    <q-modal v-if="document.fileType === 'text/csv' && fileURL !== ''" v-model="textRead" maximized>
+      <q-btn class="overlay-button fixed z-max" style="top: 20px; right: 20px;" color="primary" icon="fas fa-times" @click.native="textRead = false" />
+      <embed :src="fileURL" width="100%" type="text/plain" id="text-object">
+    </q-modal> -->
   </q-page>
 </template>
 
@@ -70,8 +107,21 @@ export default {
           value: 'viewed'
         }
       ],
+      fileTypes: [
+        'application/msword',
+        'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+        'application/vnd.ms-powerpoint',
+        'application/vnd.openxmlformats-officedocument.presentationml.presentation',
+        'application/vnd.ms-excel',
+        'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+      ],
+      pdfRead: false,
+      wordRead: false,
+      htmlRead: false,
+      textRead: false,
       fileURL: '',
-      pdfLoaded: pdf === {}
+      currentPage: 1,
+      pageCount: 0
     }
   },
   mounted () {
@@ -90,6 +140,25 @@ export default {
           // })
         })
       })
+    },
+    nextPage () {
+      if (this.currentPage !== this.pageCount) {
+        this.currentPage++
+      }
+    },
+    prevPage () {
+      if (this.currentPage !== 1) {
+        this.currentPage--
+      }
+    },
+    handleSwipe (obj) {
+      if (obj.direction === 'left') {
+        this.nextPage()
+      } else if (obj.direction === 'right') {
+        this.prevPage()
+      } else {
+        console.error('not valid swipe direction: ', obj)
+      }
     },
     updateNotes (notes) {
       this.userData.notes = notes
@@ -145,4 +214,22 @@ export default {
 </script>
 
 <style>
+
+.pdf-viewer {
+  position: relative;
+  width: 100%;
+}
+
+.overlay-button {
+  opacity: .5;
+}
+
+.overlay-button:hover {
+  opacity: 1;
+}
+
+#text-object pre {
+  color: var(--q-color-white) !important;
+}
+
 </style>